move JettyQueuedThreadPoolModule into the other module.
